removed VC++ warnings

+++ Source/LookAndFeel.cpp
@@ -32,7 +32,8 @@ void ChristmasLookAndFeel::drawRotarySlider (Graphics &g,
     // correct size, move it to the correct position and rotate it to the correct angle.
     AffineTransform transform = AffineTransform::scale (width / rotarySliderWidth, 
                                                         height / rotarySliderHeight);
-    transform = transform.translated (x, y);
+    transform = transform.translated (static_cast <float> (x), 
+                                      static_cast <float> (y));
     transform = transform.rotated (angle, centreX, centreY);
         
     // Draw the slider image with the transform applied.
